Add routes directory
