+++ src/app.module.ts
@@ -1,10 +1,10 @@
 import { Module } from "@nestjs/common";
 import { AppController } from "./app.controller";
 import { AppService } from "./app.service";
-import { CircuitsModule } from "./circuits/circuits.module";
-import { ConstructorsModule } from "./constructors/constructors.module";
-import { DriversModule } from "./drivers/drivers.module";
-import { SeasonsModule } from "./seasons/seasons.module";
+import { CircuitsModule } from "./routes/circuits/circuits.module";
+import { ConstructorsModule } from "./routes/constructors/constructors.module";
+import { DriversModule } from "./routes/drivers/drivers.module";
+import { SeasonsModule } from "./routes/seasons/seasons.module";
 
 @Module({
     imports: [CircuitsModule, DriversModule, ConstructorsModule, SeasonsModule],
