+++ src/routes/circuits/circuits.module.ts
@@ -1,7 +1,7 @@
 import { Module } from "@nestjs/common";
 import { CircuitsService } from "./circuits.service";
 import { CircuitsController } from "./circuits.controller";
-import { PrismaService } from "../prisma.service";
+import { PrismaService } from "../../prisma.service";
 
 @Module({
     controllers: [CircuitsController],
