+++ src/routes/circuits/circuits.service.test.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from "@nestjs/testing";
-import { PrismaService } from "../prisma.service";
+import { PrismaService } from "../../prisma.service";
 import { CircuitsService } from "./circuits.service";
 
 describe("CircuitsService", () => {
