+++ src/routes/circuits/dto/get-circuits.dto.ts
@@ -1,19 +1,27 @@
 import { Type } from "class-transformer";
 import { IsInt, IsNotEmpty, IsOptional, IsPositive } from "class-validator";
-import { CommonDto } from "../../common.dto";
+import { CommonDto } from "../../../common.dto";
 
-export class GetSeasonsDto extends CommonDto {
+export class GetCircuitsDto extends CommonDto {
     @IsOptional()
     @IsNotEmpty()
     driverId: string;
 
     @IsOptional()
     @IsNotEmpty()
-    circuitId: string;
-
-    @IsOptional()
-    @IsNotEmpty()
     constructorId: string;
+
+    @Type(() => Number)
+    @IsOptional()
+    @IsInt()
+    @IsPositive()
+    year: number;
+
+    @Type(() => Number)
+    @IsOptional()
+    @IsInt()
+    @IsPositive()
+    round: number;
 
     @Type(() => Number)
     @IsOptional()
@@ -38,16 +46,4 @@ export class GetSeasonsDto extends CommonDto {
     @IsInt()
     @IsPositive()
     status: number; // TODO: validate for actual status values
-
-    @Type(() => Number)
-    @IsOptional()
-    @IsInt()
-    @IsPositive()
-    driverStandings: number;
-
-    @Type(() => Number)
-    @IsOptional()
-    @IsInt()
-    @IsPositive()
-    constructorStandings: number;
 }
