+++ src/routes/constructors/constructors.module.ts
@@ -1,7 +1,7 @@
 import { Module } from "@nestjs/common";
 import { ConstructorsService } from "./constructors.service";
 import { ConstructorsController } from "./constructors.controller";
-import { PrismaService } from "../prisma.service";
+import { PrismaService } from "../../prisma.service";
 
 @Module({
     controllers: [ConstructorsController],
