+++ src/routes/drivers/drivers.module.ts
@@ -1,7 +1,7 @@
 import { Module } from "@nestjs/common";
 import { DriversService } from "./drivers.service";
 import { DriversController } from "./drivers.controller";
-import { PrismaService } from "../prisma.service";
+import { PrismaService } from "../../prisma.service";
 
 @Module({
     controllers: [DriversController],
