+++ src/routes/seasons/dto/get-seasons.dto.ts
@@ -1,27 +1,19 @@
 import { Type } from "class-transformer";
 import { IsInt, IsNotEmpty, IsOptional, IsPositive } from "class-validator";
-import { CommonDto } from "../../common.dto";
+import { CommonDto } from "../../../common.dto";
 
-export class GetCircuitsDto extends CommonDto {
+export class GetSeasonsDto extends CommonDto {
     @IsOptional()
     @IsNotEmpty()
     driverId: string;
 
     @IsOptional()
     @IsNotEmpty()
+    circuitId: string;
+
+    @IsOptional()
+    @IsNotEmpty()
     constructorId: string;
-
-    @Type(() => Number)
-    @IsOptional()
-    @IsInt()
-    @IsPositive()
-    year: number;
-
-    @Type(() => Number)
-    @IsOptional()
-    @IsInt()
-    @IsPositive()
-    round: number;
 
     @Type(() => Number)
     @IsOptional()
@@ -46,4 +38,16 @@ export class GetCircuitsDto extends CommonDto {
     @IsInt()
     @IsPositive()
     status: number; // TODO: validate for actual status values
+
+    @Type(() => Number)
+    @IsOptional()
+    @IsInt()
+    @IsPositive()
+    driverStandings: number;
+
+    @Type(() => Number)
+    @IsOptional()
+    @IsInt()
+    @IsPositive()
+    constructorStandings: number;
 }
