+++ src/routes/seasons/seasons.module.ts
@@ -1,7 +1,7 @@
 import { Module } from "@nestjs/common";
 import { SeasonsService } from "./seasons.service";
 import { SeasonsController } from "./seasons.controller";
-import { PrismaService } from "../prisma.service";
+import { PrismaService } from "../../prisma.service";
 
 @Module({
     controllers: [SeasonsController],
