+++ src/routes/seasons/types/Season.ts
